feat(server): implement base server listener
Implement basic flow structure for the creation of a site files server. A basic listener with
logging was implemented, and a placeholder pathname processor was supplied in the meantime. A
`Configuration` interface was added to `@ndcb/config` so that there is no type validation to do in
`@ndcb/server` and `@ndcb/builder`.