refactor: use string tags instead of symbols
Symbols are unique when constructed. Using symbols for discriminated unions works in settings where
there is only one "same instance" of the symbol; symbols discriminate against versions of a package.
This would not pose problems for this project. However, this would make it so that a library `lib`
using packages `pkg1` and `pkg2` where `pkg1` depends on `pkg2` would require `lib` to use the same
version of `pkg2` in order for the discriminated types to agree. Therefore, we resort back to using
strings as tags for discriminated unions.