refactor(mock-fs): simplify implementation for mocked entry existence queries
Existing mocked files necessarily have a corresponding contents string in the hash map used in the
mocked read file function. Likewise, existing mocked directories necessarily have a corresponding
array of entries in the hash map used in the mocked read directory function. Thus, it suffices to
assert that these hash maps have respectively a queried file or directory as key to return that they
exist in the mock file system.