feat(fs): implement file retrieval from pathname
A file in a file system has a pathname. In composite file systems, the file in the system of least
index takes precedence over the others. Hence with this ordering, there is only one visible file for
a given pathname. It is then possible to retrieve the file corresponding to a source pathname. This
will be useful for routing and processing.